Add files via upload
In this game, players take turns rolling a single die as many times as they wish, accumulating their roll results into a running total. However, if a player rolls a 1, they lose all points gained during that turn.

The game features three buttons: "ROLL DICE," "HOLD," and "NEW GAME."

ROLL DICE: Pressing this button rolls the die, yielding a number between 1 and 6. The number rolled is added to the current score for that turn. If a 1 is rolled, the player loses all points for that turn and next player get the chance.
HOLD: This button adds the current score to the player's overall score and ends their turn.
NEW GAME: Resets the game to start fresh.
Players can continue rolling the die and adding to their current score until they choose to "HOLD" or roll a 1. The goal is to strategically accumulate points without losing them by rolling a 1.

And if someones overall score is greater than or equal to 20 in that case that player win the game.